refactor: utility class의 성격을 지닌 Utils클래스 패키지 이동
