+++ src/main/java/racinggame/domain/RacingGame.java
@@ -6,6 +6,7 @@ import java.io.IOException;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
+import racinggame.util.Utils;
 
 public class RacingGame {
